@@ -8,63 +8,71 @@
   "dependencies": {
     "@fortawesome/fontawesome": "^1.1.8",
     "@fortawesome/fontawesome-free-brands": "^5.0.13",
-    "@fortawesome/react-fontawesome": "0.0.20",
+    "@fortawesome/fontawesome-svg-core": "^1.2.6",
+    "@fortawesome/react-fontawesome": "0.1.3",
+    "@wapps/gatsby-plugin-i18next": "^1.0.5",
+    "@wapps/gatsby-plugin-material-ui": "^1.0.7",
     "autoprefixer": "^8.5.1",
     "axios": "^0.18.0",
-    "dotenv": "^6.0.0",
-    "gatsby": "1.9.270",
-    "gatsby-image": "^1.0.5",
-    "gatsby-link": "1.6.44",
-    "gatsby-plugin-google-analytics": "^1.0.31",
-    "gatsby-plugin-google-fonts": "0.0.4",
-    "gatsby-plugin-i18n": "^0.4.2",
-    "gatsby-plugin-manifest": "^1.0.22",
-    "gatsby-plugin-offline": "^1.0.16",
-    "gatsby-plugin-postcss-sass": "^1.0.20",
-    "gatsby-plugin-react-helmet": "^2.0.11",
-    "gatsby-plugin-react-next": "1.0.11",
-    "gatsby-plugin-sharp": "^1.6.44",
-    "gatsby-plugin-sitemap": "^1.2.23",
-    "gatsby-remark-images": "^1.5.63",
-    "gatsby-source-contentful": "1.3.52",
-    "gatsby-transformer-remark": "1.7.42",
-    "gatsby-transformer-sharp": "^1.6.24",
-    "jss": "^9.8.1",
-    "lodash": "^4.16.4",
-    "react": "^16.4.0",
-    "react-async-script": "^0.9.1",
-    "react-dom": "^16.4.0",
-    "react-google-recaptcha": "^0.11.1",
+    "dotenv": "^6.1.0",
+    "gatsby": "^2.0.27",
+    "gatsby-codemods": "^1.0.5",
+    "gatsby-image": "^2.0.15",
+    "gatsby-link": "^2.0.4",
+    "gatsby-plugin-google-analytics": "^2.0.6",
+    "gatsby-plugin-jss": "^2.0.2",
+    "gatsby-plugin-manifest": "^2.0.5",
+    "gatsby-plugin-material-ui": "^0.1.3",
+    "gatsby-plugin-offline": "^2.0.7",
+    "gatsby-plugin-postcss": "^2.0.0",
+    "gatsby-plugin-purgecss": "^2.4.0",
+    "gatsby-plugin-react-helmet": "^3.0.0",
+    "gatsby-plugin-sass": "^2.0.1",
+    "gatsby-plugin-sharp": "^2.0.7",
+    "gatsby-plugin-sitemap": "^2.0.1",
+    "gatsby-remark-images": "^2.0.4",
+    "gatsby-source-contentful": "^2.0.4",
+    "gatsby-transformer-remark": "^2.1.8",
+    "gatsby-transformer-sharp": "^2.1.4",
+    "jss": "^9.8.7",
+    "lodash": "^4.17.11",
+    "node-sass": "^4.9.3",
+    "react": "^16.5.2",
+    "react-async-script": "^1.0.0",
+    "react-dom": "^16.5.2",
+    "react-event-listener": "^0.6.4",
+    "react-google-recaptcha": "^1.0.4",
     "react-helmet": "^5.2.0",
-    "react-jss": "^8.4.0",
+    "react-jss": "^8.6.1",
     "react-scrollable-anchor": "^0.6.1",
-    "react-typography": "^0.15.0",
-    "slash": "^1.0.0",
-    "smooth-scroll": "^14.2.0",
-    "typography": "^0.15.8",
-    "typography-breakpoint-constants": "^0.14.0"
+    "react-typography": "^0.16.3",
+    "slash": "^2.0.0",
+    "smooth-scroll": "^14.2.1",
+    "typography": "^0.16.17",
+    "typography-breakpoint-constants": "^0.15.10"
   },
   "devDependencies": {
-    "@material-ui/icons": "1.0.0-beta.43",
-    "animate.css": "^3.5.2",
+    "@material-ui/core": "^3.2.2",
+    "@material-ui/icons": "^3.0.1",
+    "animate.css": "^3.7.0",
     "babel-eslint": "^8.0.3",
     "devicon-2.2": "^2.2.0",
     "emergence.js": "^1.1.2",
-    "eslint": "^4.13.0",
+    "eslint": "^5.6.0",
     "eslint-config-prettier": "^2.9.0",
-    "eslint-plugin-prettier": "^2.6.0",
     "eslint-plugin-import": "^2.12.0",
     "eslint-plugin-jsx-a11y": "^6.0.3",
+    "eslint-plugin-prettier": "^2.6.0",
     "eslint-plugin-react": "^7.5.1",
     "font-awesome": "^4.7.0",
-    "gh-pages": "^1.1.0",
-    "husky": "^0.14.3",
+    "gh-pages": "^2.0.1",
+    "husky": "^1.1.2",
     "lint-staged": "^6.0.0",
     "material-ui": "1.0.0-beta.45",
-    "normalize.css": "^7.0.0",
-    "prettier": "1.13.5",
-    "pretty-quick": "^1.6.0",
-    "textlint": "^10.1.1",
+    "normalize.css": "^8.0.0",
+    "prettier": "^1.14.3",
+    "pretty-quick": "^1.8.0",
+    "textlint": "^11.0.1",
     "textlint-rule-preset-ja-spacing": "^2.0.1",
     "textlint-rule-preset-japanese": "^4.0.0"
   },
