@@ -1,5 +1,5 @@
 import React from 'react';
-import { navigateTo } from 'gatsby-link';
+import { navigate } from 'gatsby';
 import Recaptcha from 'react-google-recaptcha';
 import axios from 'axios';
 
@@ -53,7 +53,7 @@ export default class Contact extends React.Component {
       headers: { 'Content-Type': 'application/x-www-form-urlencoded' },
       body: encode({ 'form-name': 'contact', ...this.state })
     })
-      .then(() => navigateTo(form.getAttribute('action')))
+      .then(() => navigate(form.getAttribute('action')))
       .catch((error) => alert(error));
 
     e.preventDefault();
