@@ -1,50 +1,41 @@
-/* eslint-disable no-underscore-dangle */
-
-import { SheetsRegistry } from 'jss';
-import { createMuiTheme, createGenerateClassName } from 'material-ui/styles';
-// import purple from '@material-ui/core/colors/purple';
-// import green from '@material-ui/core/colors/green';
+import createMuiTheme from 'material-ui/styles/createMuiTheme'
+import purple from '@material-ui/core/colors/purple';
+import green from '@material-ui/core/colors/green';
 
 // A theme with custom primary and secondary color.
 // It's optional.
-const theme = createMuiTheme({
-  // palette: {
-  //   primary: {
-  //     light: purple[300],
-  //     main: purple[500],
-  //     dark: purple[700],
-  //   },
-  //   secondary: {
-  //     light: green[300],
-  //     main: green[500],
-  //     dark: green[700],
-  //   },
-  // },
+export default createMuiTheme({
+  typography: {
+    useNextVariants: true,
+  }
 });
 
-function createPageContext() {
-  return {
-    theme,
-    // This is needed in order to deduplicate the injection of CSS in the page.
-    sheetsManager: new Map(),
-    // This is needed in order to inject the critical CSS.
-    sheetsRegistry: new SheetsRegistry(),
-    // The standard class name generator.
-    generateClassName: createGenerateClassName()
-  };
-}
+// const createPageContext = () => {
+//   return {
+//     theme,
+//     // This is needed in order to deduplicate the injection of CSS in the page.
+//     sheetsManager: new Map(),
+//     // This is needed in order to inject the critical CSS.
+//     sheetsRegistry: new SheetsRegistry(),
+//     // The standard class name generator.
+//     generateClassName: createGenerateClassName()
+//   };
+// };
 
-export default function getPageContext() {
-  // Make sure to create a new context for every server-side request so that data
-  // isn't shared between connections (which would be bad).
-  if (!process.browser) {
-    return createPageContext();
-  }
+// const getPageContext = () => {
+//   // Make sure to create a new context for every server-side request so that data
+//   // isn't shared between connections (which would be bad).
 
-  // Reuse context on the client-side.
-  if (!global.__INIT_MATERIAL_UI__) {
-    global.__INIT_MATERIAL_UI__ = createPageContext();
-  }
+//   if (!process.browser) {
+//     return createPageContext();
+//   }
 
-  return global.__INIT_MATERIAL_UI__;
-}
+//   // Reuse context on the client-side.
+//   if (!global.__INIT_MATERIAL_UI__) {
+//     global.__INIT_MATERIAL_UI__ = createPageContext();
+//   }
+
+//   return global.__INIT_MATERIAL_UI__;
+// };
+
+// export default getPageContext;
