@@ -12,13 +12,13 @@ function withRoot(Component) {
       this.pageContext = this.props.pageContext || getPageContext();
     }
 
-    componentDidMount() {
-      // Remove the server-side injected CSS.
-      const jssStyles = document.querySelector('#server-side-jss');
-      if (jssStyles && jssStyles.parentNode) {
-        jssStyles.parentNode.removeChild(jssStyles);
-      }
-    }
+    // componentDidMount() {
+    //   // Remove the server-side injected CSS.
+    //   const jssStyles = document.querySelector('#server-side-jss');
+    //   if (jssStyles && jssStyles.parentNode) {
+    //     jssStyles.parentNode.removeChild(jssStyles);
+    //   }
+    // }
 
     pageContext = null;
 
@@ -26,16 +26,20 @@ function withRoot(Component) {
       const locale = this.props.location.pathname.includes('/en')
         ? 'en-US'
         : 'fr';
+
+        // console.log(this.pageContext)
+
       // MuiThemeProvider makes the theme available down the React tree thanks to React context.
       return (
-        <MuiThemeProvider
-          theme={this.pageContext.theme}
-          sheetsManager={this.pageContext.sheetsManager}
-        >
-          {/* CssBaseline kickstart an elegant, consistent, and simple baseline to build upon. */}
-          <CssBaseline />
-          <Component {...this.props} locale={locale} />
-        </MuiThemeProvider>
+        // <MuiThemeProvider
+        //   theme={this.pageContext.theme}
+        //   sheetsManager={this.pageContext.sheetsManager}
+        // >
+        //   {/* CssBaseline kickstart an elegant, consistent, and simple baseline to build upon. */}
+        //   <CssBaseline />
+        //   <Component {...this.props} locale={locale} />
+        // </MuiThemeProvider>
+        <Component {...this.props} locale={locale} />
       );
     }
   }
